+++ app/Main.hs
@@ -9,19 +9,30 @@ import Data.Semigroup((<>))
 import Options.Applicative
 import System.Environment (getArgs)
 
+import Options.Output (PrintOptions(..))
 import PixelTransform (geomImage)
 import Scad (scadify)
-
-data PrintOptions = PrintOptions
-  { invert :: Bool
-  , filename :: String}
 
 printOptions :: Parser PrintOptions
 printOptions = PrintOptions
                <$> switch
-               ( long "invert"
-               <> short 'i'
-               <> help "Inverts the height of each pixel, so that brighter pixels are taller and shorter are longer.")
+                  ( long "invert"
+                  <> short 'i'
+                  <> help "Inverts the height of each pixel, so that brighter pixels are taller and shorter are longer.")
+               <*> option auto
+                  ( long "length"
+                  <> short 'l'
+                  <> help "Sets the length of the output."
+                  <> metavar "LEN"
+                  <> value 100
+                  <> showDefault)
+               <*> option auto
+                  ( long "width"
+                  <> short 'w'
+                  <> help "Sets the width of output."
+                  <> metavar "WIDTH"
+                  <> value 100
+                  <> showDefault)
                <*> strArgument (metavar "FILE")
 
 printOptionsInfo :: ParserInfo PrintOptions
@@ -33,7 +44,7 @@ main :: IO ()
 main = do
   options <- execParser printOptionsInfo
   eitherImage <- readImage $ filename options
-  either (putStrLn) (putStrLn . (getScadCode $ invert options) . convertRGB8) eitherImage
+  either (putStrLn) (putStrLn . (getScadCode $ options) . convertRGB8) eitherImage
 
-getScadCode :: Bool -> Image PixelRGB8 -> String
-getScadCode b = scadify . (geomImage b)
+getScadCode :: PrintOptions -> Image PixelRGB8 -> String
+getScadCode po = scadify . (geomImage po)