+++ src/Options.hs
@@ -1,0 +1,2 @@
+module Options (PrintOptions(..)) where
+import Options.Output